requirements.txt
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
urllib3>=2.2.0
pytest>=8.0.0

# Airflow and providers
apache-airflow>=2.8.0
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-snowflake>=4.0.0
snowflake-sqlalchemy>=1.5.0
snowflake-connector-python>=3.1.0
